Refactored code by pushing down the method
